+++ lib/td_dq/search/aggregations.ex
@@ -8,15 +8,15 @@ defmodule TdDq.Search.Aggregations do
   def rule_aggregation_terms do
     static_keywords = [
       {"active.raw", %{terms: %{field: "active.raw"}}},
-      {"domain_parents",
-       %{
-         nested: %{path: "domain_parents"},
-         aggs: %{distinct_search: %{terms: %{field: "domain_parents.name.raw", size: 50}}}
-       }},
       {"current_business_concept_version",
        %{terms: %{field: "current_business_concept_version.name.raw", size: 50}}},
       {"execution_result_info.result_text",
-       %{terms: %{field: "execution_result_info.result_text.raw", size: 50}}}
+       %{terms: %{field: "execution_result_info.result_text.raw", size: 50}}},
+       {"taxonomy",
+       %{
+         nested: %{path: "domain_parents"},
+         aggs: %{distinct_search: %{terms: %{field: "domain_parents.id", size: 50}}}
+       }}
     ]
 
     ["dq", "bg"]
@@ -27,17 +27,17 @@ defmodule TdDq.Search.Aggregations do
 
   def implementation_aggregation_terms do
     static_keywords = [
-      {"domain_parents",
-       %{
-         nested: %{path: "domain_parents"},
-         aggs: %{distinct_search: %{terms: %{field: "domain_parents.name.raw", size: 50}}}
-       }},
       {"current_business_concept_version",
        %{terms: %{field: "current_business_concept_version.name.raw", size: 50}}},
       {"execution_result_info.result_text",
        %{terms: %{field: "execution_result_info.result_text.raw", size: 50}}},
       {"rule", %{terms: %{field: "rule.name.raw", size: 50}}},
-      {"source_external_id", %{terms: %{field: "structure_aliases.raw", size: 50}}}
+      {"source_external_id", %{terms: %{field: "structure_aliases.raw", size: 50}}},
+      {"taxonomy",
+       %{
+         nested: %{path: "domain_parents"},
+         aggs: %{distinct_search: %{terms: %{field: "domain_parents.id", size: 50}}}
+       }}
     ]
 
     ["dq", "bg", "ri"]