+++ lib/td_dq/search/mappings.ex
@@ -68,7 +68,10 @@ defmodule TdDq.Search.Mappings do
           errors: %{type: "long"},
           records: %{type: "long"},
           last_execution_at: %{type: "date", format: "strict_date_optional_time||epoch_millis"},
-          result_text: %{type: "text", fields: @raw}
+          result_text: %{
+            type: "text",
+            fields: %{raw: %{type: "keyword", null_value: "quality_result.no_execution"}}
+          }
         }
       },
       _confidential: %{type: "boolean"},
@@ -91,7 +94,7 @@ defmodule TdDq.Search.Mappings do
       domain_parents: %{
         type: "nested",
         properties: %{
-          id: %{type: "long", index: false},
+          id: %{type: "long"},
           name: %{type: "text", fields: @raw}
         }
       },
@@ -135,7 +138,10 @@ defmodule TdDq.Search.Mappings do
           errors: %{type: "long"},
           records: %{type: "long"},
           date: %{type: "date", format: "strict_date_optional_time||epoch_millis"},
-          result_text: %{type: "text", fields: @raw}
+          result_text: %{
+            type: "text",
+            fields: %{raw: %{type: "keyword", null_value: "quality_result.no_execution"}}
+          }
         }
       },
       _confidential: %{type: "boolean"},