+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- [TD-3867] Exception calculating some lineage graphs (graph 1.2.0)
+
 ## [4.23.0] 2021-06-28
 
 ### Fixed
@@ -29,6 +35,7 @@
   each structure, assume 1 if structure has any linked concepts and 0 otherwise.
 - [TD-3718] Get the extra information when structures are downloaded
 - [TD-3864] Issue serializing certain lineage graphs as JSON
+
 ### Added
 
 - [TD-3736] Tags in data structure version document