+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- [TD-3752] Show execution implementation filter with no result data.
 - [TD-3867] Exception calculating some lineage graphs (graph 1.2.0)
 
 ## [4.23.0] 2021-06-28