+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - [TD-3787] Allow CSV bulk load of structure notes with autopublish capability
   for non admin users.
+### Fixed
+
+- [TD-3867] Exception calculating some lineage graphs (graph 1.2.0)
 
 ## [4.23.0] 2021-06-28
 
@@ -35,6 +38,7 @@
   each structure, assume 1 if structure has any linked concepts and 0 otherwise.
 - [TD-3718] Get the extra information when structures are downloaded
 - [TD-3864] Issue serializing certain lineage graphs as JSON
+
 ### Added
 
 - [TD-3736] Tags in data structure version document