+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-3787] Allow CSV bulk load of structure notes with autopublish capability
+  for non admin users. These updates must follow new structure notes' workflow.
+  
 ### Changed
 
 - [TD-3933] Maximum size of payload for metadata upload using multipart data can