+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 ## [Unreleased]
 
+### Added
+
+- [TD-3787] Allow CSV bulk load of structure notes with autopublish capability
+  for non admin users. These updates must follow new structure notes' workflow.
+  
+### Changed
+
+- [TD-3933] Maximum size of payload for metadata upload using multipart data can
+  now be configured using the `MAX_PAYLOAD_LENGTH` environment variable
+
 ### Fixed
 
+- [TD-3752] Show execution implementation filter with no result data
 - [TD-3867] Exception calculating some lineage graphs (graph 1.2.0)
 
 ### Added